+++ RELEASENOTES.md
@@ -40,6 +40,8 @@
     *   Enable wake lock handling by default to fix issues with buffering during
         background playback. This is equivalent to setting
         `ExoPlayer.Builder.setWakeMode` to `C.WAKE_MODE_LOCAL`.
+    *   Add `ExoPlayer.setVirtualDeviceId` to update the virtual device ID
+        originally obtained from the `Context` passed to `ExoPlayer.Builder`.
     *   Ensure renderers don't consume data from the next playlist item more
         than 10 seconds before the end of the current item.
     *   Add `setSeekBackIncrementMs`, `setSeekForwardIncrementMs` and